repositorio para tecnico

+++ src/main/java/com/example/apiprojeto/repository/TecnicoRepository.java
@@ -3,5 +3,5 @@ package com.example.apiprojeto.repository;
 import com.example.apiprojeto.domain.Tecnico;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TecnicoRepository extends JpaRepository<Tecnico, Integer> {
+public interface TecnicoRepository extends JpaRepository<Tecnico, Long> {
 }
